Move HashStable implementations.
